configure/lsf: Make failure to find yp_all() more clear.
It was recently brought to our attention that building OMPI
with LSF on RHEL requires three different libnsl libraries.
Unfortunately, this isn't clear from the output of configure,
which makes it seem like a libevent problem. Make the failure
more apparent, with a link to the faq for more info.

An update to the faq will be provided to the www repo with
this information.

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>

Co-authored-by: Jeff Squyres <jsquyres@users.noreply.github.com>